Started moving hero cards to new state functions
