+++ src/components/builder-mode/BuilderMode.tsx
@@ -1,5 +1,5 @@
 import { useStore } from "../../state/store.ts";
-import { SelectionMenu } from "./SelectionMenu";
+import { SelectionMenu } from "./selection/SelectionMenu";
 import { Warbands } from "./warbands/Warbands.tsx";
 
 // TODO: Update the builder mode components to use typescript.
@@ -32,6 +32,12 @@ export const BuilderMode = () => {
     });
   };
 
+  const warbandFocus = roster.warbands.find(
+    ({ id }) => id === warriorSelectionFocus[0],
+  );
+
+  console.log(warriorSelectionFocus);
+
   return (
     <>
       <SelectionMenu
@@ -43,9 +49,7 @@ export const BuilderMode = () => {
         setFactionSelection={setFactionSelection}
         heroSelection={heroSelection}
         newWarriorFocus={warriorSelectionFocus[1]}
-        warbandNumFocus={
-          roster.warbands.find(({ id }) => id === warriorSelectionFocus[0])?.num
-        }
+        warbandNumFocus={warbandFocus?.num - 1}
       />
       <Warbands />
     </>
